Restore interruption status on InterruptedException
As described in Java Concurrency In Practice, pp. 92ff.,
InterruptedException should either be propagated to the caller
or the interruption status should be restored.
See https://cdn.rawgit.com/HackathonHackers/programming-ebooks/master/Java/Java%20Concurrency%20in%20Practice.pdf#page=114
(Go to pages 114f.. of the PDF if it does not scroll automatically, it's pages 92ff. in the original book)